# public/private-swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/private/events":{"post":{"description":"Creates an event in the database and sends it to Klaviyo. The event is saved locally regardless of Klaviyo API status.","operationId":"eventcontroller-create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventDto"},"examples":{"Active on Site":{"summary":"Example: Customer active on site","value":{"eventName":"Active on Site","eventAttributes":{"page":"/home","source":"web"},"profileAttributes":{"email":"customer@example.com","first_name":"John","last_name":"Doe","phone_number":"+1234567890"}}},"Viewed Product":{"summary":"Example: Customer viewed a product","value":{"eventName":"Viewed Product","eventAttributes":{"productId":"PROD-789","productName":"Laptop","price":1299.99},"profileAttributes":{"email":"customer@example.com","first_name":"John"}}}}}}},"responses":{"201":{"description":"Event created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventResponseDto"},"examples":{"success":{"summary":"Event created successfully","value":{"success":true,"message":"Event created successfully","data":{"id":1,"eventName":"Active on Site","createdAt":"2024-11-17T10:30:00.000Z"}}}}}}},"400":{"description":"Bad Request - Invalid input data"}},"summary":"Create a single event","tags":["Events"]}},"/api/private/events/bulk":{"post":{"description":"Creates multiple events in the database and sends them to Klaviyo in a single batch. All events are saved locally regardless of Klaviyo API status.","operationId":"eventcontroller-createbulk","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBulkEventsDto"},"examples":{"Multiple Events":{"summary":"Example: Create multiple events at once","value":{"events":[{"eventName":"Active on Site","eventAttributes":{"page":"/home","source":"web"},"profileAttributes":{"email":"customer1@example.com","first_name":"John","last_name":"Doe"}},{"eventName":"Viewed Product","eventAttributes":{"productId":"PROD-123","productName":"Laptop","price":1299.99},"profileAttributes":{"email":"customer2@example.com","first_name":"Jane","last_name":"Smith"}}]}}}}}},"responses":{"201":{"description":"Bulk events created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkEventResponseDto"},"examples":{"success":{"summary":"Bulk events created successfully","value":{"success":true,"message":"Bulk events created successfully","data":{"created":2,"failed":0,"details":[{"id":1,"eventName":"Active on Site","createdAt":"2024-11-17T10:30:00.000Z"},{"id":2,"eventName":"Viewed Product","createdAt":"2024-11-17T10:30:05.000Z"}]}}}}}}},"400":{"description":"Bad Request - Invalid input data"}},"summary":"Create multiple events in bulk","tags":["Events"]}},"/api/private/metrics":{"get":{"description":"Retrieves all available metrics from Klaviyo with pagination and search support","operationId":"metriccontroller-getallmetrics","parameters":[{"name":"page","required":false,"in":"query","description":"Page number for pagination","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page (max: 100)","schema":{"default":10,"type":"number"}},{"name":"search","required":false,"in":"query","description":"Search term for global search across event names, profile emails, and metric names","schema":{"type":"string"}},{"name":"filters","required":false,"in":"query","description":"Filters in JSON format. Available types: includesString, greaterOrEqual, lessOrEqual, equals, notEquals, arrIncludes, isDateBetween, equalDateYear. Use for filtering events by metric, profiles by email, or date ranges.","schema":{"type":"string"}},{"name":"orders","required":false,"in":"query","description":"Sorting orders in JSON format. Direction can be ASC or DESC. Common columns: createdAt, eventName, email","schema":{"type":"string"}},{"name":"start_date","required":false,"in":"query","description":"Start date for date range filtering (format: YYYY-MM-DD). Used to filter events created on or after this date.","schema":{"type":"string"}},{"name":"end_date","required":false,"in":"query","description":"End date for date range filtering (format: YYYY-MM-DD). Used to filter events created on or before this date.","schema":{"type":"string"}}],"responses":{"200":{"description":"Metrics retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MetricsResponseDto"}}}}},"summary":"Get all metrics with pagination","tags":["Metrics"]}},"/api/private/metrics/{metric}/count":{"get":{"description":"Returns the total number of events for a given metric name","operationId":"metriccontroller-getmetriccount","parameters":[{"name":"metric","required":true,"in":"path","description":"Metric name (e.g., \"Active on Site\")","schema":{"example":"Active on Site","type":"string"}}],"responses":{"200":{"description":"Event count retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MetricCountResponseDto"}}}},"404":{"description":"Metric not found"}},"summary":"Get total event count for a metric","tags":["Metrics"]}},"/api/private/metrics/{metric}/emails":{"get":{"description":"Returns a list of unique email addresses associated with events for a given metric on a specific date","operationId":"metriccontroller-getmetricemails","parameters":[{"name":"metric","required":true,"in":"path","description":"Metric name (e.g., \"Placed Order\")","schema":{"example":"Placed Order","type":"string"}},{"name":"date","required":true,"in":"query","description":"Date in YYYY-MM-DD format","schema":{"example":"2024-11-17","type":"string"}}],"responses":{"200":{"description":"Emails retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MetricEmailsResponseDto"}}}},"404":{"description":"Metric not found"}},"summary":"Get unique emails for a metric on a specific date","tags":["Metrics"]}},"/api/private/profiles/{email}/attributes":{"get":{"description":"Retrieves profile attributes from Klaviyo for a given email address","operationId":"profilecontroller-getprofileattributes","parameters":[{"name":"email","required":true,"in":"path","description":"Email address of the profile","schema":{"example":"customer@example.com","type":"string"}}],"responses":{"200":{"description":"Profile attributes retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileAttributesResponseDto"}}}},"404":{"description":"Profile not found"}},"summary":"Get profile attributes by email","tags":["Profiles"]}},"/api/private/profiles/{email}/metrics":{"get":{"description":"Retrieves all events associated with a profile from Klaviyo by email address","operationId":"profilecontroller-getprofilemetrics","parameters":[{"name":"email","required":true,"in":"path","description":"Email address of the profile","schema":{"example":"customer@example.com","type":"string"}}],"responses":{"200":{"description":"Profile metrics retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileMetricsResponseDto"}}}},"404":{"description":"Profile not found"}},"summary":"Get all events/metrics for a profile","tags":["Profiles"]}}},"info":{"title":"Klaviyo backend integration","description":"API for managing external service","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000"}],"components":{"securitySchemes":{"Apikey":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"Authorization":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateEventDto":{"type":"object","properties":{"eventName":{"type":"string","description":"Name of the event (metric name in Klaviyo)","example":"Placed Order"},"eventAttributes":{"type":"object","description":"Event-specific attributes/properties","example":{"orderId":"ORD-12345","amount":99.99,"currency":"USD","items":["Product A","Product B"]}},"profileAttributes":{"type":"object","description":"Profile attributes (must include email). Use snake_case for Klaviyo fields.","example":{"email":"customer@example.com","first_name":"John","last_name":"Doe","phone_number":"+1234567890"}}},"required":["eventName","eventAttributes","profileAttributes"]},"EventDataDto":{"type":"object","properties":{"id":{"type":"number","example":1},"eventName":{"type":"string","example":"Placed Order"},"createdAt":{"format":"date-time","type":"string","example":"2024-11-17T10:30:00.000Z"},"eventAttributes":{"type":"object","example":{"orderId":"ORD-12345"}},"profileAttributes":{"type":"object","example":{"email":"customer@example.com"}}},"required":["id","eventName","createdAt","eventAttributes","profileAttributes"]},"EventResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Event created successfully"},"data":{"$ref":"#/components/schemas/EventDataDto"}},"required":["success","message","data"]},"CreateBulkEventsDto":{"type":"object","properties":{"events":{"description":"Array of events to create","example":[{"eventName":"Placed Order","eventAttributes":{"orderId":"ORD-001","amount":99.99},"profileAttributes":{"email":"customer1@example.com","firstName":"John"}},{"eventName":"Added to Cart","eventAttributes":{"productId":"PROD-123","quantity":2},"profileAttributes":{"email":"customer2@example.com","firstName":"Jane"}}],"type":"array","items":{"$ref":"#/components/schemas/CreateEventDto"}}},"required":["events"]},"BulkEventResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"Bulk events created successfully"},"data":{"type":"object","example":{"created":2,"failed":0,"details":[]}}},"required":["success","message","data"]},"MetricDto":{"type":"object","properties":{"id":{"type":"string","example":"01HQRS7YQMC3Q4X5Z6T8W9V0AB"},"type":{"type":"string","example":"metric"},"attributes":{"type":"object","example":{"name":"Placed Order","created":"2024-01-15T10:30:00Z","updated":"2024-01-15T10:30:00Z","integration":{"object":"integration","id":"01HQRS7YQMC3Q4X5Z6T8W9V0AB","name":"API","category":"Custom"}}}},"required":["id","type","attributes"]},"MetricsResponseDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/MetricDto"}}},"required":["data"]},"MetricCountResponseDto":{"type":"object","properties":{"metricName":{"type":"string","example":"Active on Site"},"count":{"type":"number","example":42}},"required":["metricName","count"]},"MetricEmailsResponseDto":{"type":"object","properties":{"metricName":{"type":"string","example":"Placed Order"},"date":{"type":"string","example":"2024-11-17"},"emails":{"example":["customer1@example.com","customer2@example.com"],"type":"array","items":{"type":"string"}},"count":{"type":"number","example":2}},"required":["metricName","date","emails","count"]},"ProfileAttributesDto":{"type":"object","properties":{"email":{"type":"string","example":"customer@example.com"},"first_name":{"type":"string","example":"John"},"last_name":{"type":"string","example":"Doe"},"phone_number":{"type":"string","example":"+1234567890"},"properties":{"type":"object","example":{"custom_field":"custom_value"}}},"required":["email"]},"ProfileAttributesResponseDto":{"type":"object","properties":{"id":{"type":"string","example":"01HQRS7YQMC3Q4X5Z6T8W9V0AB"},"attributes":{"$ref":"#/components/schemas/ProfileAttributesDto"}},"required":["id","attributes"]},"ProfileEventDto":{"type":"object","properties":{"id":{"type":"string","example":"01HQRS7YQMC3Q4X5Z6T8W9V0AB"},"type":{"type":"string","example":"event"},"attributes":{"type":"object","example":{"timestamp":"2024-11-17T10:30:00Z","event_properties":{"orderId":"ORD-12345","amount":99.99},"datetime":"2024-11-17T10:30:00Z"}}},"required":["id","type","attributes"]},"ProfileMetricsResponseDto":{"type":"object","properties":{"email":{"type":"string","example":"customer@example.com"},"events":{"type":"array","items":{"$ref":"#/components/schemas/ProfileEventDto"}},"totalEvents":{"type":"number","example":5}},"required":["email","events","totalEvents"]}}}}